+++ src/azure/utils.ts
@@ -1,4 +1,40 @@
+/**
+ * Normalizes locations to the most prominent form of lowercase with no spaces.
+ *
+ * @param location a location such as 'East US' or 'eastus'
+ */
+export function normalizeLocation(location?: string): string | undefined {
+  if (location) {
+    return location.toLowerCase().replace(/\s+/, "");
+  }
+}
+
 const resourceGroupRegex = /\/resourceGroups\/(.*?)\//;
+
+/**
+ * Returns the `resource.name` or throws an error. Useful in situations where
+ * you must have a name, such as when calling an API that requires the resource
+ * name.
+ *
+ * @param resource anything with a name property, really
+ */
+export function resourceName(resource: { name?: string }): string {
+  if (resource.name) {
+    return resource.name;
+  } else {
+    throw new Error("Resource name is undefined!");
+  }
+}
+
+export function resourceGroupName(
+  id: string | undefined,
+  required: true,
+): string;
+
+export function resourceGroupName(
+  id: string | undefined,
+  required?: boolean,
+): string | undefined;
 
 /**
  * Extracts the resource group name from a resource id and downcases it for